```feat:``` complete authentication with hashed password
Completed the authentication with hashed password and added the register route. Used MySQL as the database inside Planetscale. Obtained the session inside page.tsx and pass it as a prop to the Navbar component. Added a fallback avatar image. Added the authOptions.ts file to handle the authentication options for NextAuth.js. And fixed some bugs.